GCP/Org-Auto-enable-APIs/function/requirements.txt
google-cloud-firestore==2.16.0
google-api-python-client==2.147.0
google-auth-httplib2==0.2.0